Updated GUI
Added getColorHex method to return corresponding hex value for color name. Also found that right-clicking the name of a widget in the sidebar and selecting "Go To Slot" brings up a dialog that allows you to select a signal and then brings you to a new slot method created in the mainwindow class.

Also, to access widget methods from mainwindow.cpp, use ui->[name]->[method]().